// src/model/cfdi.ts
export class Cfdi {
    aprovnum: string;
    aprovyear: string;
    certificado: string;
    cfdi: boolean;
    descuento: number;
    fecha: Date;
    folio: string;
    iepsRetenido: number;
    iepsTrasladado: number;
    ishTrasladado: number;
    isrRetenido: number;
    ivaRetenido: number;
    ivaTrasladado: number;
    mensaje: string;
    moneda: string;
    noCertificado: string;
    nombreEmisor: string;
    rfcEmisor: string;
    rfcReceptor: string;
    sello: string;
    serie: string;
    subtotal: number;
    tasaiva: number;
    tipoComp: string;
    total: number;
    totalDeRetencionesLocales: number;
    totalDeTrasladosLocales: number;
    totalImpuestosRetenidos: number;
    totalImpuestosTrasladados: number;
    uuid: string;
    valida: boolean;
    version: string;
}